Additional work on delete function

+++ server/routes/votes.ts
@@ -9,21 +9,41 @@ const prisma = new PrismaClient();
 voteRouter.get(`/:userId/:typeId/:type`, async (req:any, res:any) => {
 const {userId, typeId, type} = req.params;
 
-const typeCheckObj = {
-  id: +userId
+const findTypeObj = {
+  userId: +userId
 }
 
 if (type === 'comment') {
-typeCheckObj.commentId = +typeId;
+findTypeObj.commentId = +typeId;
 } else if (type === 'post') {
-  typeCheckObj.postId = +type.id;
+  findTypeObj.postId = +type.id;
 }
-
+console.log('obj', findTypeObj)
 try {
+const idForChecking = await prisma.vote.findFirst({
+  where: findTypeObj
+})
+// console.log('the checking id', idForChecking)
+if (idForChecking !== null) {
+
+  const typeCheckObj = {
+    id: idForChecking.id
+  }
+  
+  if (type === 'comment') {
+    typeCheckObj.commentId = +typeId;
+  } else if (type === 'post') {
+    typeCheckObj.postId = +type.id;
+  }
+  
+  
   const allMatches = await prisma.vote.findFirst({
     where: typeCheckObj,
   })
   res.status(200).send(allMatches);
+} else {
+  res.status(200).send('no match')
+}
 
 } catch (err) {
   console.error('unable to search', err);
@@ -77,22 +97,40 @@ voteRouter.post('/:userId/:typeId/:type', async (req: any, res: any) => {
 // UNDO A LIKE
 voteRouter.delete(`/:userId/:typeId/:type`, async (req: any, res: any) => {
   const {userId, typeId, type } = req.params;
-  const deletedVote = {
-    id: +userId
+  const findingVote = {
+    userId: +userId
   }
   
   if (type === 'comment') {
-    deletedVote.commentId = +typeId;
+    findingVote.commentId = +typeId;
   } else if (type === 'post') {
-    deletedVote.postId = +type.id;
+    findingVote.postId = +type.id;
   }
   try {
-    
-    const loseVote = await prisma.vote.delete({
-      where: deletedVote,
+    const findVote = await prisma.vote.findFirst({
+      where: findingVote,
     })
-    console.log(loseVote);
-    res.status(200).send('successful deletion');
+    if (findVote) {
+
+      const deletedVote = {
+        id: +findVote.id
+      }
+      
+      if (type === 'comment') {
+        deletedVote.commentId = +typeId;
+      } else if (type === 'post') {
+        deletedVote.postId = +type.id;
+      }
+      
+      console.log('found', findVote)
+      const loseVote = await prisma.vote.delete({
+        where: deletedVote,
+      })
+      console.log(loseVote);
+      res.status(200).send('successful deletion');
+    } else {
+      res.status(404).send('no vote found');
+    }
   } catch (err) {
     console.error('failed to delete vote', err);
   res.sendStatus(500);
